requirements.txt
alembic==0.8.2.dev0
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
Babel==1.3
bcrypt==3.1.6
blinker==1.3
Brlapi==0.6.4
certifi==2019.3.9
cffi==1.12.2
chardet==2.3.0
Click==7.0
command-not-found==0.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.3
cryptography==2.6.1
defer==1.0.6
defusedxml==0.5.0
dirspec==13.10
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-cors-headers==2.5.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-generate-scaffold==0.0.5a0
django-guardian==1.5.0
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.9.2
djangorestframework-datatables==0.4.1
djangorestframework-jwt==1.11.0
ecdsa==0.13
Fabric==1.10.2
Flask==0.10.1
Flask-BabelEx==0.9.3
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==0.6.7
Flask-Security==1.7.5
Flask-SQLAlchemy==1.0
Flask-WTF==0.14.2
htmlmin==0.1.12
httplib2==0.9.1
idna==2.0
invoke==1.2.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.8
kazam==1.4.5
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.0.1
MarkupSafe==0.23
numpy==1.15.2
oauthlib==3.0.1
onboard==1.2.0
oneconf==0.3.9
paramiko==2.4.2
passlib==1.6.5
pbr==5.1.3
pexpect==4.0.1
Pillow==3.1.2
pipenv==2018.11.26
piston-mini-client==0.7.5
psutil==3.4.2
psycopg2==2.7.6.1
ptyprocess==0.5
pyasn1==0.1.9
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.7.1
PyNaCl==1.3.0
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.4.2
python-debian==0.1.27
python-xapp==1.0.0
python3-openid==3.1.0
pytz==2018.9
pyxdg==0.25
PyYAML==5.1
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==1.2.0
sessioninstaller==0.0.0
setproctitle==1.1.8
simplejson==3.8.1
six==1.10.0
SQLAlchemy==1.0.11
sqlparse==0.2.4
sshtunnel==0.1.4
stevedore==1.30.1
system-service==0.3
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.35
uritemplate==3.0.0
urllib3==1.13.1
virtkey==0.63.0
virtualenv==16.4.3
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.4
Werkzeug==0.10.4
WTForms==2.1
xkit==0.0.0
